pyproject.toml
[tool.poetry]
name = "throttle-cli"
version = "0.1.0"
description = "Throttle commands triggered from multiple async sources."
authors = ["Bence Ferdinandy <bence@ferdinandy.com>"]
readme = "README.md"
license = "EUPL-1.2"
homepage = "https://sr.ht/~ferdinandyb/throttle/"

[tool.poetry.dependencies]
python = ">=3.10"
jsonrpclib-pelix = "^0.4.3.2"
toml = "^0.10.2"
pyxdg = "^0.28"

[tool.poetry.scripts]
throttle = "throttle_cli.cli:main"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
ruff = "^0.2.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
lint.select = ["E", "F", "I001"] # default + isort
line-length = 88 # black compatible

[tool.mypy]
ignore_missing_imports = true
